feat: Update project to remove references to vulnerable libraries (#156)
The POM and project structure of the Maven Plugin has been stagnant for years, and now transitive
dependencies of certain dependencies and plugins have started to flag the plugin as a security threat,
despite none of the vulnerabilities being applicable to the normal use case. Furthermore, some
vulnerable dependencies are banned on an organizational level for some clients. To remedy this, we
update the entire plugin to a modern Maven 3 spec. As a side effect, the plugin now requires
Maven 3.9.9+ and Java 11+ to build.